backend/chatapp/requirements.txt
django==6.0
djangorestframework==3.15.2
channels==4.1.0
daphne==4.1.2
django-cors-headers==4.5.0
Pillow==11.0.0
channels-redis==4.2.0
